feat:[section]:[client-server-actions]=> handel update and delete to skills and experiencs with server actions.

+++ client/app/actions/delete/actions.js
@@ -1,22 +1,27 @@
 "use server";
 import credentials from "@/app/credentials/credentials";
 import { revalidatePath } from "next/cache";
+import { redirect } from "next/dist/server/api-utils";
 
 export async function deleteExperience(id) {
   const { user, isLogged } = await credentials();
   try {
-    const requestDelete = await fetch(
-      `http://localhost:4000/api/${user.id}/experiences/${id}`,
-      {
-        method: "DELETE",
-        headers: {
-          "Content-Type": "application/json",
-        },
-      }
-    );
-    const data = requestDelete.json();
-    revalidatePath(`/experiences`);
-    return data;
+    if (isLogged) {
+      const requestDelete = await fetch(
+        `http://localhost:4000/api/${user.id}/experiences/${id}`,
+        {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+          },
+        }
+      );
+      const data = requestDelete.json();
+      revalidatePath(`/experiences`);
+      return data;
+    } else {
+      redirect("/api/auth/login");
+    }
   } catch (error) {
     return {
       error: "cant't delete experiences",
@@ -24,5 +29,55 @@ export async function deleteExperience(id) {
     };
   }
 }
-export async function deleteProject(fromData, id) {}
-export async function deleteSkill(fromData, id) {}
+export async function deleteProject(id) {
+  const { user, isLogged } = await credentials();
+  try {
+    if (isLogged) {
+      const requestDelete = await fetch(
+        `http://localhost:4000/api/${user.id}/project/${id}`,
+        {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+          },
+        }
+      );
+      const data = requestDelete.json();
+      revalidatePath(`/projects`);
+      return data;
+    } else {
+      redirect("/api/auth/login");
+    }
+  } catch (error) {
+    return {
+      error: "cant't delete this project",
+      message: error.message,
+    };
+  }
+}
+export async function deleteSkill(id) {
+  const { user, isLogged } = await credentials();
+  try {
+    if (isLogged) {
+      const requestDelete = await fetch(
+        `http://localhost:4000/api/${user.id}/skills/${id}`,
+        {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+          },
+        }
+      );
+      const data = requestDelete.json();
+      revalidatePath(`/skills`);
+      return data;
+    } else {
+      redirect("/api/auth/login");
+    }
+  } catch (error) {
+    return {
+      error: "cant't delete skill",
+      message: error.message,
+    };
+  }
+}
